# backend/StigviddAPI/Controllers/ReviewsController.cs
// SPDX-FileCopyrightText: 2025-2026 The Stigvidd Authors
// SPDX-License-Identifier: AGPL-3.0-or-later

using Core.Interfaces.Services;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using WebDataContracts.RequestModels.Review;
using WebDataContracts.ResponseModels.Review;

namespace StigviddAPI.Controllers;

[ApiController]
[Route("/api/v1/[controller]")]
public class ReviewsController : StigViddController
{
    private readonly IReviewService _reviewService;
    private readonly IUserService _userService;

    public ReviewsController(IReviewService service, IUserService userService)
    {
        _reviewService = service;
        _userService = userService;
    }

    [AllowAnonymous]
    [HttpGet]
    [Route("trail/{trailIdentifier}")]
    public async Task<ActionResult<PagedReviewResponse>> GetReviewsByTrailIdentifier(
        [FromRoute] string trailIdentifier,
        [FromQuery] int page, // Vilken omgång
        [FromQuery] int limit, // Hur många per omgång
        CancellationToken ctoken)
    {
        var result = await _reviewService.GetReviewsByTrailIdentifierAsync(trailIdentifier, page, limit, ctoken);

        if (!result.Success && result.Message != null)
        {
            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [Authorize]
    [HttpGet]
    [Route("trail/{trailIdentifier}/mine")]
    public async Task<ActionResult<bool>> HasReviewedTrail(
        [FromRoute] string trailIdentifier,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
        {
            return Unauthorized("User not found");
        }

        var result = await _reviewService.HasUserReviewedTrailAsync(userResponse.Identifier, trailIdentifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [Authorize]
    [HttpPost]
    [Route("create")]
    public async Task<ActionResult> CreateReview(
        [FromForm] CreateReviewRequest request,
        [FromForm] IFormFileCollection? images,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
        {
            return Unauthorized("User not found");
        }

        var result = await _reviewService.AddReviewAsync(userResponse.Identifier, request.TrailIdentifier, request.TrailReview, request.Rating, images, ctoken);

        if (!result.Success && result.Message != null)
        {
            return ToActionResult(result.Message);
        }

        if (result.Value == null)
        {
            return BadRequest("Review could not be created.");
        }

        return Created($"/api/v1/review/{result.Value.Identifier}", result.Value);
    }

    [Authorize]
    [HttpDelete]
    [Route("{reviewIdentifier}")]
    public async Task<ActionResult> DeleteReview(
        [FromRoute] string reviewIdentifier,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
        {
            return Unauthorized("User not found");
        }

        var result = await _reviewService.DeleteReviewAsync(
            reviewIdentifier,
            userResponse.Identifier,
            ctoken
        );

        if (!result.Success && result.Message != null)
        {
            return ToActionResult(result.Message);
        }

        return NoContent();
    }
}