# web/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "StigVidd",
    "description": "Stigvidd API. This is free software, licensed under the GNU Affero General Public License version 3 or later. Under section 13 of that licence you are entitled to the Corresponding Source of this running service: https://github.com/themalind/stigvidd",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/v1/Notifications/tokens": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "operationId": "Notifications_RegisterToken",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Notifications/tokens/{expoToken}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "operationId": "Notifications_DeleteToken",
        "parameters": [
          {
            "name": "expoToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Reviews/trail/{trailIdentifier}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "Reviews_GetReviewsByTrailIdentifier",
        "parameters": [
          {
            "name": "trailIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedReviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Reviews/trail/{trailIdentifier}/mine": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "Reviews_HasReviewedTrail",
        "parameters": [
          {
            "name": "trailIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Reviews/create": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "operationId": "Reviews_CreateReview",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "Rating": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "TrailReview": {
                    "type": "string",
                    "nullable": true
                  },
                  "TrailIdentifier": {
                    "type": "string",
                    "nullable": true
                  },
                  "images": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Reviews/{reviewIdentifier}": {
      "delete": {
        "tags": [
          "Reviews"
        ],
        "operationId": "Reviews_DeleteReview",
        "parameters": [
          {
            "name": "reviewIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterPushTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expoToken",
          "platform"
        ],
        "properties": {
          "expoToken": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          }
        }
      },
      "PagedReviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reviews"
        ],
        "properties": {
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "hasMore": {
            "type": "boolean"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "trailIdentifier"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "trailReview": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "format": "decimal"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userIdentifier": {
            "type": "string",
            "nullable": true
          },
          "trailIdentifier": {
            "type": "string"
          },
          "reviewImages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ReviewImageResponse"
            }
          }
        }
      },
      "ReviewImageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "imageUrl"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Skriv: Bearer {din Keycloak access token}",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}